+++ src/GafferScene/ShaderTweaks.cpp
@@ -54,6 +54,7 @@
 
 #include "fmt/format.h"
 
+#include <regex>
 #include <unordered_map>
 
 using namespace std;
@@ -179,7 +180,7 @@ void checkForCycle( const ShaderNetwork &network, const IECore::InternedString &
 	}
 }
 
-bool applyTweakInternal( ShaderNetwork *shaderNetwork, unordered_map<InternedString, IECoreScene::ShaderPtr> &modifiedShaders, const TweakPlug *tweakPlug, const ShaderNetwork *inputNetwork, const std::string &tweakLabel, const ShaderNetwork::Parameter &parameter, const IECoreScene::Shader *shader, TweakPlug::MissingMode missingMode, bool &removedConnections )
+bool applyTweakInternal( ShaderNetwork *shaderNetwork, unordered_map<InternedString, IECoreScene::ShaderPtr> &modifiedShaders, const TweakPlug *tweakPlug, const ShaderNetwork *inputNetwork, const std::string &tweakLabel, const ShaderNetwork::Parameter &parameter, const std::optional<std::string> &shaderTypeFilter, const IECoreScene::Shader *shader, TweakPlug::MissingMode missingMode, bool &removedConnections )
 {
 	if( !shader )
 	{
@@ -199,6 +200,11 @@ bool applyTweakInternal( ShaderNetwork *shaderNetwork, unordered_map<InternedStr
 		{
 			return false;
 		}
+	}
+
+	if( shaderTypeFilter && shader->getName() != *shaderTypeFilter )
+	{
+		return false;
 	}
 
 	const TweakPlug::Mode mode = static_cast<TweakPlug::Mode>( tweakPlug->modePlug()->getValue() );
@@ -382,6 +388,11 @@ bool applyTweakInternal( ShaderNetwork *shaderNetwork, unordered_map<InternedStr
 			missingMode
 		);
 	}
+}
+
+IECore::InternedString regexSubMatchToInterned( const std::ssub_match &subMatch )
+{
+	return IECore::InternedString( &( *subMatch.first ), subMatch.length() );
 }
 
 }  // namespace
@@ -591,24 +602,55 @@ bool ShaderTweaks::applyTweaks( IECoreScene::ShaderNetwork *shaderNetwork, Tweak
 
 		}
 
+		const static std::regex shaderTypeRegex( R"(([^{}]*)(\{shaderType=(.*)\})?\.(.*))" );
+
 		ShaderNetwork::Parameter parameter;
-		const size_t dotPos = name.find_last_of( '.' );
-		if( dotPos == string::npos )
+		std::optional<std::string> shaderTypeFilter;
+
+		std::smatch regexMatch;
+
+		// In order to be a valid expression for finding a shader parameter, it should match the regex, and have
+		// either a shader handle, or shaderType qualifier
+		if( std::regex_match( name, regexMatch, shaderTypeRegex ) && ( regexMatch[1].length() || regexMatch[2].length() ) )
 		{
-			parameter.shader = shaderNetwork->getOutput().shader;
-			parameter.name = name;
+			if( regexMatch[2].length() )
+			{
+				// Subgroup 3 is just the actual shader type
+				shaderTypeFilter = regexSubMatchToInterned( regexMatch[3] );
+			}
+
+			if( regexMatch[1].length() )
+			{
+				parameter.shader = regexSubMatchToInterned( regexMatch[1] );
+			}
+			else
+			{
+				// It's allowed to omit the handle when using a shaderType filter,
+				// in which case we want to match everything with the shaderType
+				parameter.shader = "*";
+			}
+			parameter.name = regexSubMatchToInterned( regexMatch[4] );
 		}
 		else
 		{
-			parameter.shader = InternedString( name.c_str(), dotPos );
-			parameter.name = InternedString( name.c_str() + dotPos + 1 );
+			// If we weren't able to parse it, the name should be a simple parameter name, which we
+			// find on the output shader. If there are any special tokens in the name at this point,
+			// something has gone wrong.
+			const static std::regex hasSyntax( R"([{}.])" );
+			if( std::regex_search( name, hasSyntax ) )
+			{
+				throw IECore::Exception( fmt::format( "Could not parse shader parameter: \"{}\"", name  ) );
+			}
+
+			parameter.shader = shaderNetwork->getOutput().shader;
+			parameter.name = name;
 		}
 
 		if( !IECore::StringAlgo::hasWildcards( parameter.shader.string() ) )
 		{
 			appliedTweaks |= applyTweakInternal(
 				shaderNetwork, modifiedShaders, tweakPlug.get(), inputNetwork,
-				name, parameter, nullptr,
+				name, parameter, shaderTypeFilter, nullptr,
 				missingMode, removedConnections
 			);
 		}
@@ -620,7 +662,7 @@ bool ShaderTweaks::applyTweaks( IECoreScene::ShaderNetwork *shaderNetwork, Tweak
 				{
 					appliedTweaks |= applyTweakInternal(
 						shaderNetwork, modifiedShaders, tweakPlug.get(), inputNetwork,
-						name, { s.first, parameter.name }, s.second.get(),
+						name, { s.first, parameter.name }, shaderTypeFilter, s.second.get(),
 						TweakPlug::MissingMode::Ignore, removedConnections
 					);
 				}